Add `collapseAxis` to `View`’s properties.
